+++ CodeGen/Generators/UnitsNetGen/UnitTestStubGenerator.cs
@@ -1,7 +1,6 @@
-using CodeGen.Helpers;
 using CodeGen.JsonTypes;
 
-namespace CodeGen.Generators
+namespace CodeGen.Generators.UnitsNetGen
 {
     internal class UnitTestStubGenerator : GeneratorBase
     {
