+++ CodeGen/Generators/UnitsNetGenerator.cs
@@ -6,14 +6,15 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
+using CodeGen.Generators.UnitsNetGen;
 using CodeGen.Helpers;
 using CodeGen.JsonTypes;
 using Newtonsoft.Json;
 using Serilog;
 
-namespace CodeGen.Generators.WindowsRuntimeComponent
+namespace CodeGen.Generators
 {
-    internal static class UnitsNetWrcGenerator
+    internal static class UnitsNetGenerator
     {
         private const int AlignPad = 35;
 
@@ -27,7 +28,6 @@ namespace CodeGen.Generators.WindowsRuntimeComponent
         {
             if (repositoryRoot == null) throw new ArgumentNullException(nameof(repositoryRoot));
             var root = repositoryRoot.FullName;
-            var outputDir = $"{root}/UnitsNet.WindowsRuntimeComponent/GeneratedCode";
 
             var templatesDir = Path.Combine(root, "Common/UnitDefinitions");
             var jsonFiles = Directory.GetFiles(templatesDir, "*.json");
@@ -36,15 +36,17 @@ namespace CodeGen.Generators.WindowsRuntimeComponent
             foreach (Quantity quantity in quantities)
             {
                 var sb = new StringBuilder($"{quantity.Name}:".PadRight(AlignPad));
-                GenerateQuantity(sb, quantity, $"{outputDir}/Quantities/{quantity.Name}.WindowsRuntimeComponent.g.cs");
-                GenerateUnitType(sb, quantity, $"{outputDir}/Units/{quantity.Name}Unit.g.cs");
+                GenerateQuantity(sb, quantity, $"{root}/UnitsNet/GeneratedCode/Quantities/{quantity.Name}.NetFramework.g.cs"); // TODO Remove NetFramework suffix
+                GenerateUnitType(sb, quantity, $"{root}/UnitsNet/GeneratedCode/Units/{quantity.Name}Unit.g.cs");
+                GenerateUnitTestBaseClass(sb, quantity, $"{root}/UnitsNet.Tests/GeneratedCode/{quantity.Name}TestsBase.g.cs");
+                GenerateUnitTestClassIfNotExists(sb, quantity, $"{root}/UnitsNet.Tests/CustomCode/{quantity.Name}Tests.cs");
                 Log.Information(sb.ToString());
             }
 
             Log.Information("");
-            GenerateUnitAbbreviationsCache(quantities, $"{outputDir}/UnitAbbreviationsCache.g.cs");
-            GenerateQuantityType(quantities, $"{outputDir}/QuantityType.g.cs");
-            GenerateStaticQuantity(quantities, $"{outputDir}/Quantity.g.cs");
+            GenerateUnitAbbreviationsCache(quantities, $"{root}/UnitsNet/GeneratedCode/UnitAbbreviationsCache.g.cs");
+            GenerateQuantityType(quantities, $"{root}/UnitsNet/GeneratedCode/QuantityType.g.cs");
+            GenerateStaticQuantity(quantities, $"{root}/UnitsNet/GeneratedCode/Quantity.g.cs");
 
             var unitCount = quantities.SelectMany(q => q.Units).Count();
             Log.Information("");
