+++ CodeGen/Generators/UnitsNetWrcGen/QuantityTypeGenerator.cs
@@ -1,7 +1,6 @@
-using CodeGen.Helpers;
 using CodeGen.JsonTypes;
 
-namespace CodeGen.Generators
+namespace CodeGen.Generators.UnitsNetWrcGen
 {
     internal class QuantityTypeGenerator : GeneratorBase
     {
@@ -27,15 +26,11 @@ namespace UnitsNet
     /// </summary>
     public enum QuantityType
     {
-// Missing XML comment for public type or member
-#pragma warning disable CS1591
         Undefined = 0,");
             foreach (var quantity in _quantities)
                 Writer.WL($@"
         {quantity.Name},");
             Writer.WL(@"
-// Missing XML comment for public type or member
-#pragma warning restore CS1591
     }
 }");
             return Writer.ToString();
