+++ CodeGen/Program.cs
@@ -4,7 +4,6 @@ using System.IO;
 using System.Linq;
 using System.Reflection;
 using CodeGen.Generators;
-using CodeGen.Generators.WindowsRuntimeComponent;
 using Serilog;
 using Serilog.Events;
 
@@ -44,7 +43,7 @@ namespace CodeGen
                     Log.Debug($"verbose: {true}", ConsoleColor.Blue);
                 }
 
-                Generator.Generate(repositoryRoot);
+                UnitsNetGenerator.Generate(repositoryRoot);
                 UnitsNetWrcGenerator.Generate(repositoryRoot);
                 Log.Information($"Completed in {sw.ElapsedMilliseconds} ms!", ConsoleColor.Green);
                 return 0;
